+++ src/types/index.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    TDI: typeof TDI;
+  }
+}
+declare var __DEV__: boolean;
+declare var TDI: ITDI;