+++ tests/responses/script1.js
@@ -1,5 +1,0 @@
-/*QUnit.test( 'TDI.Ajax.Response: external script', function( assert ) {
-	assert.expect(1);
-	
-	assert.ok( true, 'External script loaded.' );
-} );*/