+++ tests/responses/script2.js
@@ -1,1 +1,0 @@
-window._TDI_TEST_SCRIPT_EXTERNAL_2 = true;